fix mixin issue in fabric module

+++ neoforge/src/main/kotlin/top/katton/api/event/ServerLivingEntityEvent.kt
@@ -45,7 +45,7 @@ object ServerLivingEntityEvent {
         onLivingFall(
             LivingFallArg(
                 e.entity,
-                e.distance.toFloat(),
+                e.distance,
                 e.damageMultiplier
             )
         )
